Set IsMouseVisible = false; added load/save config buttons

+++ MorpheusInTheUnderworld/Classes/GameSettings.cs
@@ -25,10 +25,15 @@ namespace MorpheusInTheUnderworld.Classes
         // This method will write to a configuration.txt file.
         public static void Write()
         {
-            string[] lines = { "master_volume:" + MasterVolume,
+            if (File.Exists(configurationPath) && ConfirmOverwrite())
+            {
+
+
+                string[] lines = { "master_volume:" + MasterVolume,
                                "music_volume:" + MusicVolume,
                                "effects_volume:" + EffectsVolume};
-            File.WriteAllLines(configurationPath, lines);
+                File.WriteAllLines(configurationPath, lines);
+            }
         }
 
         // this method will read from a configuration.txt file
@@ -46,6 +51,10 @@ namespace MorpheusInTheUnderworld.Classes
                 MusicVolume = Convert.ToInt32(lines[1].Split(':')[1]);
                 EffectsVolume = Convert.ToInt32(lines[2].Split(':')[1]);
             }
+        }
+        public static bool ConfirmOverwrite()
+        {
+            return true;
         }
     }
 }
